Make Ors better for objects
When joing together 2 converters with the or method the error
message has traditionally been terrible, and reported that the expected value
was either object. This
replaces the dumb logic with a hueristic to select the most errors
at the lowest depth to determine which branch should
have passed. It's not perfect but it's better
than what was there.