Run converters before None checks
In the current world, the optional decorators will return immediately
if the field matches the optional type. This means that chained
converters such as path, or, etc, will be ignored. This commit changes
the order of operations so the converter will be run before trying to
return the optional value.